backend/requirements.txt
fastapi[all]
tensorflow
beautifulsoup4
Pillow